--- apps/frontend/frontend-2/.eslintrc.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../.eslintrc.json",
-  "ignorePatterns": ["!**/*"],
-  "overrides": [
-    {
-      "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
-      "parserOptions": { "project": ["apps/frontend-2/tsconfig.*?.json"] },
-      "rules": {}
-    },
-    { "files": ["*.ts", "*.tsx"], "rules": {} },
-    { "files": ["*.js", "*.jsx"], "rules": {} }
-  ]
-}
