--- apps/frontend/frontend-2/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  displayName: 'frontend-2',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsConfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/frontend-2',
-};
