--- apps/frontend/frontend-2/src/components/Footer.svelte
@@ -1,4 +1,0 @@
-<div class="flex flex-col min-w-full py-2 bg-gray-100">
-    <span class="font-light text-center">Note: This is a demo</span>
-    <span class="font-light text-center">© Blockchain Acceleration Foundation 2021</span>
-</div>
