--- apps/frontend/frontend-2/src/components/base/Card.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-    export let classOverride: string | undefined;
-    export let classExtra: string | string[] = ""
-
-    const extraClasses = Array.isArray(classExtra) ? classExtra.join(" ") : classExtra;
-    const classes = classOverride !== undefined ? classOverride : `p-4 bg-gray-50 rounded-md shadow-md ${extraClasses}`
-</script>
-
-<div class={classes}>
-    <slot></slot>
-</div>
