--- apps/frontend/frontend-2/src/config/constants.ts
@@ -1,4 +1,0 @@
-import { Envs } from '@baf-wallet/interfaces';
-export const constants = {
-  env: Envs.DEV,
-};
