--- apps/frontend/frontend-2/src/main.ts
@@ -1,10 +1,0 @@
-import App from './App.svelte';
-
-const app = new App({
-  target: document.body,
-  props: {
-    name: 'frontend',
-  },
-});
-
-export default app;
