--- apps/frontend/frontend-2/src/pages/Login.svelte
@@ -1,4 +1,0 @@
-<script lang="ts">
-import Auth from "../components/Auth.svelte";
-</script>
-<Auth />
