--- apps/frontend/frontend-2/src/pages/NotFound404.svelte
@@ -1,2 +1,0 @@
-404 ;(
-The page you are looking for is not found
