--- apps/frontend/frontend-2/src/state/keys.svelte
@@ -1,6 +1,0 @@
-<script lang="ts" context="module">
-  import { writable } from 'svelte/store';
-  import { KeyState } from '../interfaces';
-
-  export const KeyStore = writable<KeyState | null>(null);
-</script>
