--- apps/frontend/frontend-2/tailwind.config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  purge: {
-    content: ['./src/**/*.svelte'],
-    enabled: process.env.production, // disable purge in dev
-  },
-  darkMode: false, // or 'media' or 'class'
-  theme: {
-    extend: {},
-  },
-  variants: {
-    extend: {},
-  },
-  plugins: [
-    require('@tailwindcss/typography'),
-    require('@tailwindcss/forms'),
-    require('@tailwindcss/line-clamp'),
-    require('@tailwindcss/aspect-ratio'),
-  ],
-};
