--- apps/frontend/frontend-2/tsconfig.json
@@ -1,25 +1,0 @@
-{
-	"include": ["src/**/*", "webpack.config.ts"],
-	"exclude": ["node_modules/*", "__sapper__/*", "public/*"],
-  "extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"rootDir": "../../",
-		"moduleResolution": "node",
-		"target": "es2017",
-		"importsNotUsedAsValues": "remove",
-		"isolatedModules": true,
-		"esModuleInterop": true,
-		"sourceMap": true,
-		"strict": false,
-		"baseUrl": "../../",
-		"lib": ["DOM"],
-		"types": [
-			"svelte",
-			"svelte/store",
-			"svelte/motion",
-			"svelte/transition",
-			"svelte/animate",
-			"svelte/easing"
-		]
-	}
-}
