--- apps/frontend/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "frontend",
-  "version": "0.0.1",
-  "dependencies": {}
-}
