+++ apps/frontend/public/index.html
@@ -1,16 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width,initial-scale=1">
 
-    <title>Svelte app</title>
+		<title>Svelte app</title>
 
-    <link rel="icon" type="image/png" href="/favicon.png" />
-    <link rel="stylesheet" href="/bundle.css" />
+		<link rel="icon" type="image/png" href="/favicon.png">
 
-    <script defer src="/bundle.js"></script>
-  </head>
-
-  <body></body>
+		<link rel="stylesheet" href="/build/bundle.css">
+		<script defer src="/build/bundle.js"></script>
+	</head>
+	<body>
+		<noscript>Please enable JavaScript to view this application.</noscript>
+	</body>
 </html>
