+++ apps/frontend/src/state/keys.svelte
@@ -1,6 +1,6 @@
 <script lang="ts" context="module">
   import { writable } from 'svelte/store';
-  import type { KeyState } from '../interfaces';
+  import { KeyState } from '../interfaces';
 
   export const KeyStore = writable<KeyState | null>(null);
 </script>
