+++ apps/frontend/tsconfig.json
@@ -1,29 +1,25 @@
 {
+	"include": ["src/**/*", "webpack.config.ts"],
+	"exclude": ["node_modules/*", "__sapper__/*", "public/*"],
   "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "target": "es2017",
-    "importsNotUsedAsValues": "error",
-    "isolatedModules": true,
-    "sourceMap": true,
-    "types": ["node", "svelte"],
-    "strict": false,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "lib": ["ES2017", "DOM"],
-    "forceConsistentCasingInFileNames": true
-  },
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.app.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    }
-  ]
+	"compilerOptions": {
+		"rootDir": "../../",
+		"moduleResolution": "node",
+		"target": "es2017",
+		"importsNotUsedAsValues": "remove",
+		"isolatedModules": true,
+		"esModuleInterop": true,
+		"sourceMap": true,
+		"strict": false,
+		"baseUrl": "../../",
+		"lib": ["DOM"],
+		"types": [
+			"svelte",
+			"svelte/store",
+			"svelte/motion",
+			"svelte/transition",
+			"svelte/animate",
+			"svelte/easing"
+		]
+	}
 }
